fix(run): align bin fallback paths
